# openapi/src/eds/ccip/eds-ccip.yaml
openapi: 3.0.0
info:
  title: CCIP Explicit Disclosure API
  description: Provides access to explicit disclosures required to send/execute CCIP messages from/to Canton.
  version: 0.0.1

paths:
  /ccip/v1/global/PerPartyRouter/factory:
    post:
      operationId: postPerPartyRouterFactory
      tags:
        - general
      description: |
        Get the explicit disclosures for instantiating a PerPartyRouter using the PerPartyRouterFactory.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CCIPPerPartyRouterFactoryRequest'
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CCIPPerPartyRouterFactoryResponse'
        '400':
          $ref: '../common/eds-common.yaml#/components/responses/400'
        '404':
          $ref: '../common/eds-common.yaml#/components/responses/404'
        '500':
          $ref: '../common/eds-common.yaml#/components/responses/500'

  /ccip/v1/global/TokenAdminRegistry/token/{instrumentId}:
    get:
      operationId: getTokenAdminRegistryToken
      tags:
        - general
      description: |
        Get the token pool for a given instrument from the token admin registry.
      parameters:
        - in: path
          name: instrumentId
          required: true
          schema:
            $ref: '../common/eds-common.yaml#/components/schemas/HashedInstrumentId'
          description: The InstrumentId of the asset, as a hashed value, to look up in the token admin registry.
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LookupTokenPoolResponse'
        '400':
          $ref: '../common/eds-common.yaml#/components/responses/400'
        '404':
          $ref: '../common/eds-common.yaml#/components/responses/404'
        '500':
          $ref: '../common/eds-common.yaml#/components/responses/500'

  /ccip/v1/global/message/send:
    post:
      operationId: postCCIPSend
      tags:
        - sending
      description: |
        Get the explicit disclosure for sending a CCIP message from Canton.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CCIPSendRequest'
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CCIPSendResponse"
        '400':
          $ref: '../common/eds-common.yaml#/components/responses/400'
        '404':
          $ref: '../common/eds-common.yaml#/components/responses/404'
        '500':
          $ref: '../common/eds-common.yaml#/components/responses/500'

  /ccip/v1/global/message/execute:
    post:
      operationId: postCCIPExecute
      tags:
        - execution
      description: |
        Get the explicit disclosures for executing a CCIP message on Canton.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CCIPExecuteRequest'
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CCIPExecuteResponse"
        '400':
          $ref: '../common/eds-common.yaml#/components/responses/400'
        '404':
          $ref: '../common/eds-common.yaml#/components/responses/404'
        '500':
          $ref: '../common/eds-common.yaml#/components/responses/500'

tags:
  - name: general
    description: General endpoints
  - name: sending
    description: Endpoints related to sending CCIP messages
  - name: execution
    description: Endpoints related to executing CCIP messages

components:
  schemas:
    CCIPPerPartyRouterFactoryRequest:
      type: object
      required:
        - partyID
      properties:
        partyID:
          $ref: "../common/eds-common.yaml#/components/schemas/PartyId"
    CCIPPerPartyRouterFactoryResponse:
      type: object
      required:
        - contractId
        - disclosedContracts
      properties:
        contractId:
          $ref: '../common/eds-common.yaml#/components/schemas/ContractId'
        disclosedContracts:
          description: Disclosures for all contracts required for the caller to instantiate a PerPartyRouter
          type: array
          items:
            $ref: '../common/eds-common.yaml#/components/schemas/DisclosedContract'

    LookupTokenPoolResponse:
      type: object
      required:
        - rawInstanceAddress
      properties:
        rawInstanceAddress:
          $ref: "../common/eds-common.yaml#/components/schemas/RawInstanceAddress"

    CCIPSendRequest:
      type: object
      required:
        - message
      properties:
        message:
          $ref: "../common/eds-common.yaml#/components/schemas/Message"
        senderRequiredCCVs:
          type: array
          items:
            $ref: "../common/eds-common.yaml#/components/schemas/RawOrHashedAddress"
        tokenPoolRequiredCCVs:
          type: array
          items:
            $ref: "../common/eds-common.yaml#/components/schemas/RawOrHashedAddress"
    CCIPSendResponse:
      type: object
      required:
        - contextData
        - disclosedContracts
        - ccvs
      properties:
        contextData:
          type: object
          description: The context to be passed along to CCIP.
        disclosedContracts:
          type: array
          items:
            $ref: '../common/eds-common.yaml#/components/schemas/DisclosedContract'
        ccvs:
          type: array
          description: |
            A list of requested CCVs containing their addresses.
            If default CCVs apply, this list might contain more CCVs than originally requested.
          items:
            $ref: "../common/eds-common.yaml#/components/schemas/RawOrHashedAddress"
          example:
            - "committeeverifier-abcde@ccipOwner::1220131b86f50178e6f7aabe62ef79e79b4964a526a1c557776409f448cbcd2406ce"
            - "0x4bed5af0e77b4d7a552a1f933b13c275bbdc231b8499b2e14edcc634a6869a5f"
        executor:
          $ref: "../common/eds-common.yaml#/components/schemas/RawOrHashedAddress"

    CCIPExecuteRequest:
      type: object
      required:
        - encodedMessage
      properties:
        encodedMessage:
          type: string
          description: The CCIP message to be executed, encoded as a hex string.
    CCIPExecuteResponse:
      type: object
      required:
        - contextData
        - disclosedContracts
      properties:
        contextData:
          type: object
          description: The context to be passed along to CCIP.
        disclosedContracts:
          type: array
          items:
            $ref: '../common/eds-common.yaml#/components/schemas/DisclosedContract'
        tokenPool:
          $ref: "../common/eds-common.yaml#/components/schemas/RawInstanceAddress"